requirements.txt
pandas>=2.2.0
psycopg2-binary>=2.9.9
SQLAlchemy>=2.0.30
python-dotenv>=1.0.1
apache-airflow>=2.9.0
fastapi>=0.111.0
uvicorn>=0.30.0
xgboost>=2.0.3
